Rename: general -> common
