+++ src/game/ai/entity/player/common/gunShoot.ts
