+++ src/game/ai/entity/player/fluff/fluffAI.ts
@@ -4,7 +4,7 @@ import { suspendable } from '@core/behaviour/suspendable'
 import { Entity } from '@core/ecs/entity'
 import { World } from '@core/ecs/world'
 import { assert } from '@utils/assertion'
-import { gunShoot } from '../general/gunShoot'
+import { gunShoot } from '../common/gunShoot'
 import { release } from './release'
 import { PLAYER_SETTING } from '../playerAI'
 import { chase } from './chase'
