+++ src/game/ai/entity/player/normal/normalAI.ts
@@ -4,7 +4,7 @@ import { Behaviour } from '@core/behaviour/behaviour'
 import { parallelAll } from '@core/behaviour/composite'
 import { animate } from '../../common/action/animate'
 import { invincibleTime } from '../../common/action/invincibleTime'
-import { gunShoot } from '../general/gunShoot'
+import { gunShoot } from '../common/gunShoot'
 import { jet } from './jet'
 import { jump } from './jump'
 import { land } from './land'
